requirements.txt
numpy
beautifulsoup4
contractions
spacy
nltk
tensorflow
keras
flask
mysql.connector